+++ .gitignore
@@ -1,4 +1,3 @@
 .env
 node_modules/
 build
-.eslintcache
