--- .prettierrc.json
@@ -1,4 +1,0 @@
-{
-  "trailingComma": "es5",
-  "singleQuote": true
-}
