+++ package.json
@@ -12,8 +12,7 @@
   },
   "dependencies": {
     "discord.js": "^12.4.1",
-    "dotenv": "^8.2.0",
-    "slugify": "^1.4.6"
+    "dotenv": "^8.2.0"
   },
   "devDependencies": {
     "@types/dotenv": "^8.2.0",
@@ -40,7 +39,7 @@
     }
   },
   "lint-staged": {
-    "*.ts": "eslint --cache --fix",
-    "*.{ts,json,md}": "prettier --write"
+    "*.js": "eslint --cache --fix",
+    "*.{js,css,md}": "prettier --write"
   }
 }
