+++ src/env.d.ts
@@ -4,8 +4,5 @@ declare namespace NodeJS {
     FORTIES_GUILD: string;
     FORTIES_SHOWCASE: string;
     FORTIES_SOUNDTEST: string;
-    IC_GB_REQUEST_CHANNEL: string;
-    IC_GB_REVIEW_CHANNEL: string;
-    IC_GB_ANNOUNCE_CHANNEL: string;
   }
 }
